requirements.txt
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docker==7.1.0
easyocr==1.7.2
faiss-cpu==1.9.0
filelock==3.16.1
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.139
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.3
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
openai==1.54.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.0.0
plotly==5.24.1
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
puremagic==1.28
pyarrow==18.0.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.13
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.40.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
traits==6.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
yarl==1.17.1
